[EGD-2571] Clenup: removed MessageType static_casts, fixed debug.h defines. Added GPIO event

+++ module-apps/messages/AppMessage.hpp
@@ -23,8 +23,8 @@ namespace app {
  */
 class AppMessage: public sys::DataMessage {
 public:
-	AppMessage( MessageType messageType ) :	sys::DataMessage( static_cast<uint32_t>(messageType)) {};
-	virtual ~AppMessage() {};
+  AppMessage(MessageType messageType) : sys::DataMessage(messageType){};
+  virtual ~AppMessage(){};
 
 };
